requirements.txt
Flask==2.3.3
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.3.0
python-dotenv==1.0.0
twilio==8.10.0
pymongo==4.5.0
python-dateutil==2.8.2
dnspython==2.4.2